# backend/src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Documentation",
    "version": "1.0.0",
    "description": "API Super Tienda",
    "contact": {
      "name": "Super    Tienda",
      "url": "https://www.superTienda.com",
      "email": "superTienda@gmail.com"
    }
  },
  "components": {
    "schemas": {
...
      "Role": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "isActive": {
            "type": "boolean",
            "default": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "users": {
            "type": "array",
            "items": {
              "$ref": "#/definitions/User"
            }
          },
          "permissions": {
            "type": "array",
            "items": {
              "$ref": "#/definitions/RolePermission"
            }
          }
        },
        "required": [
          "code",
          "name"
        ]
      },
...
      "RolePermission": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "roleId": {
            "type": "integer"
          },
          "permissionId": {
            "type": "integer"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "role": {
            "$ref": "#/definitions/Role"
          },
          "permission": {
            "$ref": "#/definitions/Permission"
          }
        },
        "required": [
          "roleId",
          "permissionId"
        ]
      },
...
      "PurchaseOrder": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "businessId": {
            "type": "integer",
            "default": 1
          },
          "supplierId": {
            "type": "integer"
          },
          "poNumber": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "string",
            "default": "DRAFT",
            "enum": [
              "DRAFT",
              "PENDING",
              "APPROVED",
              "ORDERED",
              "PARTIALLY_RECEIVED",
              "RECEIVED",
              "CANCELLED"
            ]
          },
          "orderDate": {
            "type": "string",
            "format": "date-time"
          },
          "expectedDate": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "receivedDate": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "subtotal": {
            "type": "number",
            "default": 0
          },
          "taxAmount": {
            "type": "number",
            "default": 0
          },
          "totalAmount": {
            "type": "number",
            "default": 0
          },
          "notes": {
            "type": [
              "string",
              "null"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "deletedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "createdBy": {
            "type": "integer",
            "default": 0
          },
          "updatedBy": {
            "type": "integer",
            "default": 0
          },
          "deletedBy": {
            "type": [
              "integer",
              "null"
            ]
          },
          "approvedBy": {
            "type": [
              "integer",
              "null"
            ]
          },
          "business": {
            "$ref": "#/definitions/Business"
          },
          "supplier": {
            "$ref": "#/definitions/Supplier"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/definitions/PurchaseOrderItem"
            }
          },
          "supplierDebts": {
            "type": "array",
            "items": {
              "$ref": "#/definitions/SupplierDebt"
            }
          },
          "User": {
            "anyOf": [
              {
                "$ref": "#/definitions/User"
              },
              {
                "type": "null"
              }
            ]
          },
          "userId": {
            "type": [
              "integer",
              "null"
            ]
          }
        },
        "required": [
          "supplierId"
        ]
      },
      "PurchaseOrderItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "purchaseOrderId": {
            "type": "integer"
          },
          "productId": {
            "type": "integer"
          },
          "quantity": {
            "type": "integer"
          },
          "unitCost": {
            "type": "number"
          },
          "receivedQuantity": {
            "type": "integer",
            "default": 0
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "purchaseOrder": {
            "$ref": "#/definitions/PurchaseOrder"
          },
          "product": {
            "$ref": "#/definitions/Product"
          }
        },
        "required": [
          "purchaseOrderId",
          "productId",
          "quantity",
          "unitCost"
        ]
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/purchase-order-items": {
      "get": {
        "summary": "Listar todos los ítems de órdenes de compra",
        "tags": [
          "PurchaseOrderItems"
        ],
        "responses": {
          "200": {
            "description": "Lista de ítems"
          }
        }
      },
      "post": {
        "summary": "Crear nuevo ítem en orden de compra",
        "tags": [
          "PurchaseOrderItems"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PurchaseOrderItem"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Ítem creado"
          }
        }
      }
    },
    "/api/purchase-order-items/{id}": {
      "get": {
        "summary": "Obtener ítem de orden de compra por ID",
        "tags": [
          "PurchaseOrderItems"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ítem encontrado"
          },
          "404": {
            "description": "No encontrado"
          }
        }
      },
      "put": {
        "summary": "Actualizar ítem de orden de compra",
        "tags": [
          "PurchaseOrderItems"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PurchaseOrderItem"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ítem actualizado"
          }
        }
      },
      "delete": {
        "summary": "Eliminar ítem de orden de compra",
        "tags": [
          "PurchaseOrderItems"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Eliminado exitosamente"
          }
        }
      }
    },
    "/api/purchase-orders": {
      "get": {
        "summary": "Listar todas las órdenes de compra",
        "tags": [
          "PurchaseOrders"
        ],
        "responses": {
          "200": {
            "description": "Lista de órdenes de compra"
          }
        }
      },
      "post": {
        "summary": "Crear nueva orden de compra",
        "tags": [
          "PurchaseOrders"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PurchaseOrder"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Orden de compra creada"
          }
        }
      }
    },
    "/api/purchase-orders/{id}": {
      "get": {
        "summary": "Obtener orden de compra por ID",
        "tags": [
          "PurchaseOrders"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Orden de compra encontrada"
          },
          "404": {
            "description": "No encontrada"
          }
        }
      },
      "put": {
        "summary": "Actualizar orden de compra",
        "tags": [
          "PurchaseOrders"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PurchaseOrder"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Orden de compra actualizada"
          }
        }
      },
      "delete": {
        "summary": "Eliminar orden de compra",
        "tags": [
          "PurchaseOrders"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Orden eliminada exitosamente"
          }
        }
      }
    },
    "/api/role-permissions": {
      "get": {
        "summary": "Listar todas las asignaciones de permisos a roles",
        "tags": [
          "RolePermissions"
        ],
        "responses": {
          "200": {
            "description": "Lista de asignaciones",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RolePermission"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Crear una nueva asignación rol-permiso",
        "tags": [
          "RolePermissions"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "roleId",
                  "permissionId"
                ],
                "properties": {
                  "roleId": {
                    "type": "integer",
                    "example": 1
                  },
                  "permissionId": {
                    "type": "integer",
                    "example": 3
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Asignación creada"
          }
        }
      }
    },
    "/api/role-permissions/{id}": {
      "get": {
        "summary": "Obtener una asignación por ID",
        "tags": [
          "RolePermissions"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Asignación encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RolePermission"
                }
              }
            }
          },
          "404": {
            "description": "No encontrada"
          }
        }
      },
      "put": {
        "summary": "Actualizar una asignación rol-permiso",
        "tags": [
          "RolePermissions"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "roleId",
                  "permissionId"
                ],
                "properties": {
                  "roleId": {
                    "type": "integer"
                  },
                  "permissionId": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Asignación actualizada"
          }
        }
      },
      "delete": {
        "summary": "Eliminar una asignación rol-permiso",
        "tags": [
          "RolePermissions"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Eliminado correctamente"
          }
        }
      }
    },
    "/api/roles": {
      "get": {
        "summary": "Obtener todos los roles",
        "tags": [
          "Roles"
        ],
        "responses": {
          "200": {
            "description": "Lista de roles"
          }
        }
      },
      "post": {
        "summary": "Crear un nuevo rol",
        "tags": [
          "Roles"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Role"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Rol creado"
          }
        }
      }
    },
    "/api/roles/{id}": {
      "get": {
        "summary": "Obtener un rol por ID",
        "tags": [
          "Roles"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Rol encontrado"
          }
        }
      },
      "put": {
        "summary": "Actualizar un rol",
        "tags": [
          "Roles"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Role"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Rol actualizado"
          }
        }
      },
      "delete": {
        "summary": "Eliminar un rol",
        "tags": [
          "Roles"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Eliminado"
          }
        }
      }
    },
...
  },
...
}